fix: vérification de l'existence d'un fichier avant suppression par shred

+++ .bin/scripts/setup-local-env.sh
@@ -4,7 +4,9 @@ set -euo pipefail
 echo "Updating local server/.env & ui/.env"
 
 delete_cleartext() {
-  shred -f -n 10 -u "${ROOT_DIR}/.vault_pwd.txt"
+  if [ -f "${ROOT_DIR}/.vault_pwd.txt" ]; then
+    shred -f -n 10 -u "${ROOT_DIR}/.vault_pwd.txt"
+  fi
 }
 
 trap delete_cleartext EXIT
